refactor: rename script to "modernize"
