+++ packages/app/src/commands/modernize/alternative_language.template
