+++ packages/app/src/commands/modernize/checkRequirements.js
