+++ packages/app/src/commands/modernize/createNewTranslationFiles.js
