+++ packages/app/src/commands/modernize/deleteLegacyFiles.js
