+++ packages/app/src/commands/modernize/generateTranslationMappings.js
