+++ packages/app/src/commands/modernize/main_language.template
