+++ packages/app/src/commands/modernize/splitTranslation.js
